split ComponentProvider by usage
